link sorta works

+++ src/LionFire.Trading/Accounts/PositionsExtensions.cs
@@ -9,7 +9,7 @@ namespace LionFire.Trading
     {
         public static double GetNetVolume(this Positions positions)
         {
-            long volume = 0;
+            double volume = 0;
             foreach (var p in positions)
             {
                 if (p.TradeType == TradeType.Buy)
